c308: extract UnmixingTab cluster (asNum + UnmixingTab + UnmixingSpectraCard + UnmixingTopicHeatmap + UnmixingBestMatchTable) to own file (#441 2.1)
Workspace.tsx 5191 LOC -> 4828 LOC (-363 LOC). Entry chunk 133.37 KB -> 123.60 KB (-9.8 KB).
26 tabs extracted + lazy. typecheck clean, 44/44 tests pass.
